scraping/py_requirements.txt
ipykernel
bs4
requests
function-pipes
numpy
networkx
matplotlib
supabase
python-dotenv